feat: 문장 스키마 작성 (#145)
